requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
asyncpg==0.29.0
bcrypt==4.1.2
build==1.2.1
CacheControl==0.14.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
cryptography==42.0.5
decorator==5.1.1
distlib==0.3.8
dulwich==0.21.7
executing==2.0.1
fastapi==0.109.2
fastapi-pagination==0.12.22
fastjsonschema==2.19.1
filelock==3.13.3
graphql-core==3.2.3
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
installer==0.7.0
ipython==8.23.0
jaraco.classes==3.4.0
jedi==0.19.1
jeepney==0.8.0
keyring==24.3.1
libcst==1.3.1
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.2.0
msgpack==1.0.8
packaging==24.0
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pkginfo==1.10.0
platformdirs==4.2.0
pluggy==1.4.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
pyproject_hooks==1.0.0
pytest==8.1.1
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rapidfuzz==3.8.0
redis==5.0.3
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
SecretStorage==3.3.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.36.3
strawberry-graphql==0.219.2
tomlkit==0.12.4
traitlets==5.14.2
trove-classifiers==2024.3.25
typer==0.12.3
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.27.1
uvloop==0.19.0
virtualenv==20.25.1
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
